console: Rename sidebar container folder
