+++ pkg/webui/components/sidebar/side-footer/index.js
@@ -20,7 +20,7 @@ import Dropdown from '@ttn-lw/components/dropdown'
 
 import { LanguageContext } from '@ttn-lw/lib/components/with-locale'
 
-import SidebarContext from '@console/containers/side-bar/context'
+import SidebarContext from '@console/containers/sidebar/context'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 import PropTypes from '@ttn-lw/lib/prop-types'
