+++ pkg/webui/console/containers/sidebar/header/index.js
