+++ pkg/webui/console/containers/sidebar/navigation/app-list-side-navigation.js
