+++ pkg/webui/console/containers/sidebar/navigation/app-side-navigation.js
