+++ pkg/webui/console/containers/sidebar/navigation/general-side-navigation.js
