+++ pkg/webui/console/containers/sidebar/navigation/gtw-list-side-navigation.js
