+++ pkg/webui/console/containers/sidebar/navigation/gtw-side-navigation.js
