+++ pkg/webui/console/containers/sidebar/navigation/index.js
