+++ pkg/webui/console/containers/sidebar/sidebar.styl
