+++ pkg/webui/console/containers/sidebar/story.js
@@ -23,9 +23,9 @@ import SideHeader from '@ttn-lw/components/sidebar/side-header'
 import SearchButton from '@ttn-lw/components/sidebar/search-button'
 import SideFooter from '@ttn-lw/components/sidebar/side-footer'
 
-import SidebarContext from '@console/containers/side-bar/context'
+import SidebarContext from '@console/containers/sidebar/context'
 
-import style from './side-bar.styl'
+import style from './sidebar.styl'
 
 export default {
   title: 'Sidebar/Sidebar',
